refactor: 페이지에서만 사용되는 hook page폴더로 이동
